Create Mean,Min,Max Idea
Suggested bulk code for calculating the mean, min & max values of each of the columns in the data set as set out in the course videos. Idea also included lines to calculate a total mean, min & max values but this will be developed later due to errors occurring. Idea also to attempt to limit the numbers in the output of the mean calculation as the numbers were quite large. The "data/iris1.csv" pulls the dataset from where it is stored on my personal computer with the file name iris1.csv.