Create Histograms.py
firstcol etc uses the same code to divide the columns. Each of the plots to follow show the details of each of the specified columns of data from the dataset represented on a histogram. "xlabel" pieces of code were used to show labels so that the data represented on each histogram could be identified by any user looking. The outputs have been compared to many similar analysis completed which appear online and give similar results, such as those printed at https://machinelearningmastery.com/machine-learning-in-python-step-by-step/.