Create Meanhist.py
Each histogram can also be shown individually if required by running the code in ipython separately rather than as the complete .py file, as long as the relevant libraries have been called for before the code is ran. The code was also altered to add the Column detail as a title rather than an x axis label similar to that used in the Histograms.py file.