Create DescribeTest1.py
First method discovered to use the .describe data frame function to summarize the data from the Iris data set. As this code was unsuccessful. Other methods were reviewed to enable this function which will be posted in another commit. This avenue was explored as the .describe function appeared to give a cleaner analysis for viewing purposes for information such as the count, mean, standard deviation, minimum, 25%, 50%, 75% & max from the data set than the number of lines of code required to obtain the same values for the mean, min & max in a previous commit.