Create Describe2.py
This code was successfully ran to give two different outputs depending on whether the Iris csv file loaded contained the attribute headings in the first row or not as outlined in the comments in the code. As explained in the previous describe commit, this method was used in order to obtain a cleaner viewing method in a lesser amount of code for attributes such as the minimum, maximum & mean, as the code for calculating these values for each column took more lines of code in a previous commit. The two different outputs were retrieved from the different Iris files as one only contained numeric values (iris1) while the code read the alternate file as a string due to the attribute headings.