Create Suggest.py
This code would need to be written in full to cover the 4 columns of data in the data set. With the addition of a scatter plot containing the information of the 4 columns at the end. Only suggested code at this time as I have been working on a function which would run the code for the 4 columns of data without having to write the above repetitively for each column as this would be very bulky & hard to follow.