Create Suggest1.py
Code to be repeated in one file for each  column. I would still like to add a couple of graphs but I have also been working on this code in a separate commit file. The ideal outcome would be able to call a function to run the calculations for all 4 columns of data in the data set to cut down on the amount of code required.